+++ http4k-format-jackson/src/test/kotlin/org/http4k/format/JacksonTest.kt
@@ -9,6 +9,7 @@ import org.http4k.core.Body
 import org.http4k.core.Response
 import org.http4k.core.Status.Companion.OK
 import org.http4k.core.with
+import org.http4k.format.Jackson.asA
 import org.http4k.format.Jackson.auto
 import org.http4k.format.Jackson.autoView
 import org.http4k.hamkrest.hasBody
@@ -36,6 +37,13 @@ class JacksonAutoTest : AutoMarshallingContract(Jackson) {
         val obj = ArbObject("hello", ArbObject("world", null, listOf(1), true), emptyList(), false)
 
         assertThat(body(Response(OK).with(body of listOf(obj))), equalTo(listOf(obj)))
+    }
+
+    @Test
+    fun `roundtrip list of arbitary objects to and from node`() {
+        val obj = ArbObject("hello", ArbObject("world", null, listOf(1), true), emptyList(), false)
+
+        assertThat(Jackson.asJsonObject(listOf(obj)).asA(), equalTo(listOf(obj)))
     }
 
     @Test